TIP-843: Move validator to enrichment
