+++ src/Akeneo/Pim/Enrichment/Component/Product/Validator/Constraints/ProductFilterData.php
@@ -1,27 +1,27 @@
 <?php
 
-namespace Pim\Component\Connector\Validator\Constraints;
+namespace Akeneo\Pim\Enrichment\Component\Validator\Constraints;
 
 use Symfony\Component\Validator\Constraint;
 
 /**
- * Constraint for product export filter locales structure.
- * Filter structure are "filters" for exported columns: scope, locales & attributes.
+ * Constraint for product export filter data.
+ * Filter data are Product Query Builder filters.
  *
  * @author    Philippe Mossière <philippe.mossiere@akeneo.com>
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class FilterStructureLocale extends Constraint
+class ProductFilterData extends Constraint
 {
     /** @var string */
-    public $message = 'The locale %localeCode% is not valid.';
+    public $message = 'invalid_filter_data';
 
     /**
      * {@inheritdoc}
      */
     public function validatedBy()
     {
-        return 'filter_structure_locale_validator';
+        return 'product_filter_data_validator';
     }
 }
