+++ src/Pim/Bundle/ConnectorBundle/DependencyInjection/PimConnectorExtension.php
@@ -35,7 +35,6 @@ class PimConnectorExtension extends Extension
         $loader->load('readers.yml');
         $loader->load('security.yml');
         $loader->load('steps.yml');
-        $loader->load('validators.yml');
         $loader->load('writers.yml');
         $loader->load('event_listeners.yml');
     }
